module6/Translator/requirements.txt
Flask==1.1.2
google-cloud-vision==1.0.0
google-cloud-translate==2.0.1
google-cloud-storage==1.31.1